requirements.txt
Django==2.0.5
pytz==2018.4
psycopg2-binary==2.7.4
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-extensions==2.0.7
requests==2.18.4
runscript==0.2.6
Pillow==5.1.0
gunicorn
dj-database-url